- Students detail finished

+++ app/src/main/java/com/amalip/teachers/data/dto/SetGradeRequest.kt
@@ -7,7 +7,7 @@ data class SetGradeRequest(
 
     val userId: Int = 0,
     val courseId: Int = 0,
-    val grade: Int = 0,
+    val grade: Double = 0.0,
     val partial: Int = 0
 
 )
