+++ app/src/main/java/com/amalip/teachers/domain/model/Course.kt
@@ -13,6 +13,6 @@ data class Course(
     val description: String = "",
     val picture: String = "",
     var scheduleList: MutableList<String> = mutableListOf(),
-    var grades: MutableList<Int> = mutableListOf()
+    var grades: MutableList<Double> = mutableListOf()
 
 ) : Parcelable
